+++ PointLightCollectionD3D11.cpp
@@ -11,7 +11,7 @@ PointLightCollectionD3D11::~PointLightCollectionD3D11()
 
 bool PointLightCollectionD3D11::Initialize(ID3D11Device *device, const PointLightData &lightInfo)
 {
-	const UINT lightCount = lightInfo.perLightInfo.size();
+	const UINT lightCount = static_cast<UINT>(lightInfo.perLightInfo.size());
 	_bufferData.reserve(lightCount);
 
 	for (UINT i = 0; i < lightCount; i++)
@@ -60,7 +60,7 @@ bool PointLightCollectionD3D11::Initialize(ID3D11Device *device, const PointLigh
 
 bool PointLightCollectionD3D11::UpdateLightBuffers(ID3D11DeviceContext *context) const
 {
-	const UINT lightCount = _bufferData.size();
+	const UINT lightCount = static_cast<UINT>(_bufferData.size());
 	for (UINT i = 0; i < lightCount; i++)
 	{
 		const LightBuffer *lightBuffer = &_bufferData.at(i);
@@ -78,7 +78,7 @@ bool PointLightCollectionD3D11::UpdateLightBuffers(ID3D11DeviceContext *context)
 
 UINT PointLightCollectionD3D11::GetNrOfLights() const
 {
-	return _bufferData.size();
+	return static_cast<UINT>(_bufferData.size());
 }
 
 ID3D11DepthStencilView *PointLightCollectionD3D11::GetShadowMapDSV(const UINT lightIndex) const
